# openapi.yaml
{"openapi":"3.0.1","info":{"title":"Beautiful Code Api","description":"Beautiful Code(At least I hope it is :-))","contact":{"name":"Егор","email":"goro4@mail.ru"},"version":"0.0.1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/update":{"post":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Запрос на  изменение одной точки","description":"На вход ждет Points","operationId":"updatePoints","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Points"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointResponse"}}}}}}},"/api/point":{"post":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Создание точки на карте","description":"Получение данных для создание заявки. Ждет на вход координаты и описание проблемы и опционально файл","operationId":"handleFileUpload","parameters":[{"name":"pointCoordinates","in":"query","required":true,"schema":{"type":"string"}},{"name":"description","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointResponse"}}}}}}},"/api/petition":{"post":{"tags":["Создание и обработка обращений.(изменение)"],"summary":"Создание обращения","description":"Получение данных для создание обращения. Ждет на вход тема, описание проблемы и опционально фото","operationId":"handleFileUpload_1","parameters":[{"name":"topic","in":"query","required":true,"schema":{"type":"string"}},{"name":"description","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PetitionResponse"}}}}}}},"/api/news/great":{"post":{"tags":["Создание новостей.(изменение)"],"summary":"Создание обращения","description":"Получение данных для создание обращения. Ждет на вход  описание проблемы и опционально фото","operationId":"handleFileUpload_2","parameters":[{"name":"description","in":"query","required":true,"schema":{"type":"string"}},{"name":"start","in":"query","required":true,"schema":{"type":"string"}},{"name":"end","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NewsResponse"}}}}}}},"/api/{name}":{"get":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Запрос на получение фото обращения","description":"На вход ждет имя файла","operationId":"getImage","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}},"/api/news/allNews":{"get":{"tags":["Создание новостей.(изменение)"],"summary":"Запрос на получение всех новостей","description":"опционально лимит сколько всего обращений и офсет с какого обращения","operationId":"allPointResponse","parameters":[{"name":"limit","in":"query","required":false,"schema":{"$ref":"#/components/schemas/AllNewsResponse"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AllNewsResponse"}}}}}}},"/api/image":{"get":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Запрос на получение фото обращения","description":"На вход ждет имя файла","operationId":"getImageForFront","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"image/jpeg":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}},"/api/config":{"get":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Получение версии","description":"Выдает версию ","operationId":"getConfigFront","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConfigResponse"}}}}}}},"/api/all":{"get":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Запрос на получение всех обращений","description":"На вход ждет координаты пользователя, далее опционально лимит сколько всего обращений и офсет с какого обращения","operationId":"allPointResponse_1","parameters":[{"name":"coordinates","in":"query","required":true,"schema":{"$ref":"#/components/schemas/AllPointResponse"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AllPointResponse"}}}}}}},"/api/allPetition":{"get":{"tags":["Создание и обработка обращений.(изменение)"],"summary":"Запрос на получение всех обращений","description":"опционально лимит сколько всего обращений и офсет с какого обращения","operationId":"allPointResponse_2","parameters":[{"name":"limit","in":"query","required":false,"schema":{"$ref":"#/components/schemas/AllPetitionResponse"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AllPetitionResponse"}}}}}}},"/api/{id}":{"delete":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Запрос на удаление одной точки","description":"На вход ждет id точки","operationId":"deletePoint","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointResponse"}}}}}}},"/api/news/{id}":{"delete":{"tags":["Создание новостей.(изменение)"],"summary":"Запрос на удаление одной новости","description":"На вход ждет id новости","operationId":"deletePoint_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NewsResponse"}}}}}}}},"components":{"schemas":{"Points":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"pointCoordinates":{"type":"string"},"description":{"type":"string"},"urlImage":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"PointResponse":{"type":"object","properties":{"id":{"type":"integer","description":"id из бд","format":"int64"},"status":{"type":"string","description":"статус обращения"},"coordinates":{"type":"array","description":"координаты в формате 60.497874,56.926760 ","items":{"type":"string","description":"координаты в формате 60.497874,56.926760 "}},"description":{"type":"string","description":"Описание проблемы"},"createdAt":{"type":"string","description":"Время создание/изменения обращение","format":"date-time"},"pathToImage":{"type":"string","description":"имя файла сохраненного на сервере"}},"description":"Ответ API включает в себя статус обращения,координаты,описание проблемы, дата создания обращения и имя файла"},"PetitionResponse":{"type":"object","properties":{"id":{"type":"integer","description":"id из бд","format":"int64"},"status":{"type":"string","description":"статус обращения"},"description":{"type":"string","description":"Описание проблемы"},"createdAt":{"type":"string","description":"Время создание/изменения обращение","format":"date-time"},"pathToImage":{"type":"string","description":"имя файла сохраненного на сервере"}},"description":"Ответ API включает в себя статус обращения, описание проблемы, дата создания обращения и имя файла"},"NewsResponse":{"type":"object","properties":{"id":{"type":"integer","description":"id из бд","format":"int64"},"description":{"type":"string","description":"Текст новости"},"start":{"type":"string","description":"Начало "},"end":{"type":"string","description":"Окончание"},"createdAt":{"type":"string","description":"Время создание/изменения новости","format":"date-time"},"pathToImage":{"type":"string","description":"имя файла сохраненного на сервере"}},"description":"Ответ API включает в себя статус обращения, описание проблемы, дата создания обращения и имя файла"},"AllNewsResponse":{"type":"object","properties":{"newsList":{"type":"array","description":"Список всех новостей.","items":{"$ref":"#/components/schemas/NewsResponse"}},"limit":{"type":"integer","description":"Сколько всего новостей надо.","format":"int32"},"offset":{"type":"integer","description":"С какого новостей считать.","format":"int32"}},"description":"Ответ API включает в себя Список всех новостей, лимит и офсет"},"ConfigResponse":{"type":"object","properties":{"version":{"type":"string","description":"номер версии"}},"description":"отдает номер версии"},"AllPointResponse":{"type":"object","properties":{"points":{"type":"array","description":"Список всех обращений.","items":{"$ref":"#/components/schemas/PointResponse"}},"limit":{"type":"integer","description":"Сколько всего обращений надо.","format":"int32"},"offset":{"type":"integer","description":"С какого обращения считать.","format":"int32"}},"description":"Ответ API включает в себя Список всех обращений, лимит и офсет"},"AllPetitionResponse":{"type":"object","properties":{"petitionList":{"type":"array","description":"Список всех обращений.","items":{"$ref":"#/components/schemas/PetitionResponse"}},"limit":{"type":"integer","description":"Сколько всего новостей надо.","format":"int32"},"offset":{"type":"integer","description":"С какого новостей считать.","format":"int32"}},"description":"Ответ API включает в себя Список всех обращений, лимит и офсет"}}}}